requirements.txt
flask==3.1.3
python-dotenv==1.2.2
chromadb==1.5.2
openai==2.26.0
google-genai==1.66.0
waitress==3.0.0
gunicorn==23.0.0
PyYAML==6.0.1
pytest==8.0.0
authlib==1.6.9
edge-tts==7.2.7
faster-whisper==1.2.1
flask-apispec==0.11.4
redis==7.3.0
prometheus-client==0.24.1
psutil==5.9.4
flask-talisman==1.0.0
bcrypt==4.1.2
